Rename according to naming convention from Sim
