+++ src/map/Road.py
@@ -4,7 +4,7 @@ import math
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "..", ".."))
 
-from src.map.coordinates import Coordinates
+from src.map.Coordinates import Coordinates
 
 
 class Road(object):
